requirements.txt
click
requests
transformers
torch
python-dotenv